qcommon: signed var right shift is undefined behavior

+++ src/qcommon/q_unicode.c
@@ -339,7 +339,7 @@ static void shiftbitsright(unsigned char *p, unsigned long num, unsigned long by
 			*p = 0;
 		}
 
-		*p &= (~0x00) >> num;
+		*p &= ((unsigned long)~0x00) >> num;
 
 		return;
 	}
